refactor: split dynamic truncator facade
