+++ src/shared/token-limit-truncator.ts
@@ -1,0 +1,75 @@
+import type { TruncationResult } from "./dynamic-truncator-types"
+
+const CHARS_PER_TOKEN_ESTIMATE = 4;
+
+function estimateTokens(text: string): number {
+	return Math.ceil(text.length / CHARS_PER_TOKEN_ESTIMATE);
+}
+
+export function truncateToTokenLimit(
+	output: string,
+	maxTokens: number,
+	preserveHeaderLines = 3,
+): TruncationResult {
+	if (typeof output !== 'string') {
+		return { result: String(output ?? ''), truncated: false };
+	}
+
+	const currentTokens = estimateTokens(output);
+
+	if (currentTokens <= maxTokens) {
+		return { result: output, truncated: false };
+	}
+
+	const lines = output.split("\n");
+
+	if (lines.length <= preserveHeaderLines) {
+		const maxChars = maxTokens * CHARS_PER_TOKEN_ESTIMATE;
+		return {
+			result:
+				output.slice(0, maxChars) +
+				"\n\n[Output truncated due to context window limit]",
+			truncated: true,
+		};
+	}
+
+	const headerLines = lines.slice(0, preserveHeaderLines);
+	const contentLines = lines.slice(preserveHeaderLines);
+
+	const headerText = headerLines.join("\n");
+	const headerTokens = estimateTokens(headerText);
+	const truncationMessageTokens = 50;
+	const availableTokens = maxTokens - headerTokens - truncationMessageTokens;
+
+	if (availableTokens <= 0) {
+		return {
+			result:
+				headerText + "\n\n[Content truncated due to context window limit]",
+			truncated: true,
+			removedCount: contentLines.length,
+		};
+	}
+
+	const resultLines: string[] = [];
+	let currentTokenCount = 0;
+
+	for (const line of contentLines) {
+		const lineTokens = estimateTokens(line + "\n");
+		if (currentTokenCount + lineTokens > availableTokens) {
+			break;
+		}
+		resultLines.push(line);
+		currentTokenCount += lineTokens;
+	}
+
+	const truncatedContent = [...headerLines, ...resultLines].join("\n");
+	const removedCount = contentLines.length - resultLines.length;
+
+	return {
+		result:
+			truncatedContent +
+			`\n\n[${removedCount} more lines truncated due to context window limit]`,
+		truncated: true,
+		removedCount,
+	};
+}
